Consolidated and streamlined controllers to better match managers
